fix(lint): resolve ruff 0.16.0 violations breaking CI, pin ruff upper bound
main's CI lint job has been failing on the last two pushes. Root cause:
CI's lint job (LegionForge/dev-rig's lint.yml) installs dev deps via
`pip install -e ".[dev]"`, which does NOT read this repo's uv.lock --
so the unbounded `ruff>=0.15` dependency silently resolved to 0.16.0
(latest on PyPI) instead of the 0.15.21 this repo's uv.lock pins,
while every local `uv run ruff check` stayed green and showed nothing
wrong. Reproduced by installing 0.16.0 in an isolated venv and running
CI's exact `ruff check src/convobox` invocation -- 21 errors, matching
CI's failure exactly.

Fixes, applied with review rather than blind --fix:
- Auto-fixable (ruff --fix): RET501, UP041 (asyncio.TimeoutError is a
  TimeoutError alias since Python 3.11, matches this project's >=3.12
  floor), RUF022 (__all__ sorting), I001, PIE808, B033 (a harmless
  literal duplicate in confirmword/detector.py's affirmations set --
  sets dedupe at runtime regardless, verified zero behavior change).
- Manual, each reviewed in context:
  - aec.py's SIM115: a justified noqa, not a restructure -- the wave
    handle is written to incrementally across the object's whole
    lifetime and closed explicitly in close(), not a single-function
    open/use/close that `with` could wrap.
  - capture.py's PYI034: MicrophoneStream.__enter__ return type ->
    Self.
  - playback.py's RUF046: int(round(...)) -> round(...) (round() with
    one arg already returns int).
  - playback.py's BLE001: justified noqa matching this codebase's
    existing convention (transcriber.py's _memory_diagnostic) --
    device queries fail unpredictably across drivers/host-APIs.
  - transcriber.py's G201: logger.error(..., exc_info=True) ->
    logger.exception(...), called directly inside the except block
    they're equivalent; no test asserts the literal logging call.
  - transcriber.py's SIM102: combined nested if with `and`, moved the
    explanatory comment above the combined condition.
  - orchestrator.py: ruff 0.16.0 itself flagged the existing
    `# noqa: BLE001` there as RUF100 unused -- removed on ruff's own
    say-so, not a judgment call; the prose comment explaining the
    broad catch (protecting _events_task from a hook exception)
    stays.

Also pinned `ruff>=0.15,<0.17` (was unbounded) so the next ruff
release can't silently repeat this -- CI will fail loudly on the
version bump instead of drifting quietly, giving a chance to review
the diff before raising the bound.

Also fixed one unrelated, pre-existing, real ruff finding outside
CI's src/convobox-only lint scope: an unused `pytest` import in
tests/test_approval_explanation_rendering.py (verified genuinely
unused -- no pytest.* reference anywhere in the file).

Verified: `uv run pytest` (880 passed, 1 skipped), `ruff check .`
clean on BOTH 0.15.21 (uv.lock) and 0.16.0 (CI's actual version) via
a separate isolated venv, mypy clean (pre-existing types-PyYAML stub
gap unrelated to this fix), bandit clean.

+++ src/convobox/audio/playback.py
@@ -27,7 +27,7 @@ def _resample(audio: np.ndarray, src_rate: int, dst_rate: int) -> np.ndarray:
     """
     if src_rate == dst_rate or len(audio) == 0:
         return audio
-    n_dst = int(round(len(audio) * dst_rate / src_rate))
+    n_dst = round(len(audio) * dst_rate / src_rate)
     if n_dst <= 0:
         return np.zeros(0, dtype=np.float32)
     src_x = np.linspace(0.0, 1.0, num=len(audio), endpoint=False)
@@ -99,7 +99,7 @@ def _device_output_rate(sd: object, device: str | int | None, source_rate: int) 
     try:
         info = sd.query_devices(device, "output")  # type: ignore[attr-defined]
         native = int(info["default_samplerate"])
-    except Exception:
+    except Exception:  # noqa: BLE001 -- unpredictable across drivers/host-APIs, degrade to source_rate
         return source_rate
     return native if native > 0 else source_rate
